Baseline

```
space-object (good) x 350,581 ops/sec ±0.44% (87 runs sampled)
space-object (bad) x 310,928 ops/sec ±0.51% (90 runs sampled)
```

---

1° commit: Internal: optimize pushAll

Result: slightly better performance for invalid input

```
space-object (good) x 351,413 ops/sec ±0.41% (89 runs sampled)
space-object (bad) x 330,903 ops/sec ±0.75% (90 runs sampled)
```

---

2° commit: Internal: remove `_loop_`s from output js

Result: **big performance improvement** (good catch @sledorze !)

```
space-object (good) x 514,274 ops/sec ±0.45% (88 runs sampled)
space-object (bad) x 479,642 ops/sec ±0.54% (86 runs sampled)
```

---

4° commit: Internal: flatten `chain`s

Result: good performance improvement

```
space-object (good) x 592,606 ops/sec ±0.52% (87 runs sampled)
space-object (bad) x 532,306 ops/sec ±1.18% (81 runs sampled)
```