@gcanti 2018-02-21
@sledorze working on a separate PR for the `flatten-chain` optimization