@sledorze 2018-02-22
@gcanti I have something that goes from 650 k/s to 1.1 M/s.
It is in a prototype shape; It involves a breaking change as it changes the structure of error reporting.
Not much time left but may try to finish it up.