@gcanti 2018-02-22
@sledorze Interesting. Can we release this in the meanwhile? The optimizations of this PR are backward compatible