@sledorze 2018-02-22
@gcanti I've created a new branch on that matter and opened another issue ( https://github.com/gcanti/io-ts/issues/136 ).